tempo_real/pyproject.toml
[project]
name = "tempo-real"
version = "0.1.0"
description = ""
authors = [
    {name = "Fasri",email = "felipeasr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.29.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pydrive (>=1.3.1,<2.0.0)",
    "pydrive2 (>=1.21.3,<2.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-api-python-client (>=2.161.0,<3.0.0)",
    "streamlit (>=1.42.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
